requirements-jarvis.txt
# JARVIS Core Dependencies
websockets>=12.0
watchdog>=4.0.0
pyyaml>=6.0
aiohttp>=3.9  # for Telegram API
python-telegram-bot>=20.0  # Telegram bot integration
openai>=1.0  # LLM integration (OpenAI + compatible endpoints like NVIDIA)

# Optional: embedding generation for pattern recognition
sentence-transformers>=2.2
numpy>=1.24